Separate demo content
