+++ Demo/Sources/StoriesView.swift
@@ -50,7 +50,7 @@ private struct TimeProgress: View {
 
 // Behavior: h-exp, v-exp
 struct StoriesView: View {
-    @StateObject private var model = StoriesViewModel(stories: Story.stories(from: URLPlaylist.videos))
+    @StateObject private var model = StoriesViewModel(stories: Story.stories(from: URLTemplates.videos))
 
     var body: some View {
         TabView(selection: $model.currentStory) {
